Update patch set 5
Patch Set 5:

(2 comments)

This is a first cut at the appmgr robot interface and tests.  feel free to comment.

still a bit of functionality to add (particularly allowing the user to specify a list of xapps to test, and blocking on deployed xapps actually coming ready) but it's mostly there.

Patch-set: 5