FIX: move some end-to-end test configs to values.yaml for nanobot
A few of the configurations for the ete testsuite were baked into
the testsuite file.  This change moves those configs to
the global properties file and allows them to be configured out of
the nanobot values.yaml.  This change only affects the nanobot
version of those files.

The change also makes a minor update to elminate the requirement
for a host path for logging; that is now optional.

signed-off-by: dave kormann <dk3239@att.com>
Change-Id: I02be793f3a5a6f7f51a7ec081ad372aead1f66f9