Update patch set 1
Patch Set 1: Code-Review-1

Hm.  So I'd like to see a bit more documentation about what you're trying to accomplish with this change — it looks like it largely does what the existing implementation does, but at the cost of not supporting things that aren't deployments (which is probably a nonissue for us, so maybe the simplification is a good thing.)

So, could you provide a bit of doco on the how/why of this?

Patch-set: 1
Reviewer: Gerrit User 1000047 <1000047@84e4e190-e623-4c37-9eda-58b1937c1b53>
Label: Code-Review=-1