Update patch set 2
Patch Set 2:

It seems to me that a lot of the added files here should be built into the docker image (in fact, I thought they _were_.) I've generally tried to mostly put stuff that needs helm template expansion in the configmaps (though I haven't been exclusive about that).  Could you give some sense of the rationale for this?

Also, it'd be really good to have some detail in the commit message -- not just the comments on the change -- about what's being accomplished with this change.  See previous commits for some idea (although you don't need to be nearly as verbose as I tend to be).  At minimum, the title line of the commit needs to describe why you're making the change.

Patch-set: 2
Reviewer: Gerrit User 1000047 <1000047@84e4e190-e623-4c37-9eda-58b1937c1b53>