Update patch set 8
Patch Set 8: Code-Review+1

Thanks for the updating the change.  I Think I'm good with this merging now (the verify job failure doesn't seem to be a result of this change) but I'll defer to Bharath to do the merge.

As I said, it may be necessary to bump the docker container version to get the rebuild to happen with the fixed KubernetesEntity.

Patch-set: 8
Reviewer: Gerrit User 1000047 <1000047@84e4e190-e623-4c37-9eda-58b1937c1b53>
Label: Code-Review=+1