Update patch set 10
Patch Set 10:

> Patch Set 9: Code-Review-1
> 
> (2 comments)
> 
> see comments on the dockerfile.  please use the build layer for building stuff and only copy what's necessary into the final container.

Line 59
this stuff should all happen in the build container above and the output objects copied into the runtime container to keep container size down.  Move all of this (lines 57-78 at least) up there.

>>>We have moved those lines 57 to 78 to up position and committed.

Line 65
is kubectl actually necessary in the container?  the point of the kubernetesentity library is that we use the kubernetes API and avoid exec'ing kubectl.  please eliminate this if you can.

>>>>We got confirmation from Bharath to proceed with kubectl way for parsing the log between E2sim to xapp flow. Also we showed demo to Bharath and confirmed before the code commit. 
This we done for Cherry release and for next release we are planning to add API for capturing and processing the packet between the xapp and E2sim.

Patch-set: 10