Update tests for deployment manager query params
The 'deploymentManagerId' query parameter is intended to be a UUID
formatted value since it corresponds to the attribute with the same
name in the response object (i.e., DeploymentManagerInfo).  To test
validation and usage of this parameter, it is best to split it into
two separate tests -- one to test validation of the format and another
to test the response to an unknown value.

Issue-ID: INT-165
Change-Id: I2177f6b737cfeb652dbc19bed16e1f7888b1feb5
Signed-off-by: Allain Legacy <alegacy@redhat.com>