Fix AlarmEventRecord attribute definitions
The interface spec defines several attributes using a suffix
of "ID" rather than "Id".  This includes the following attributes:

* resourceTypeID
* resourceID
* probableCauseID
* alarmDefinitionID

The document also defines perceivedSeverity as an integer field
rather than a string.

The following fields are only conditionally mandatory if alarm
corresponds to the condition being represented.  That is,
alarmChangedTime is only meant to be present if the alarm has
been modified, and alarmAcknowledgedTime is only meant to be
present if the alarm has been acknowledged.

Issued-ID: INT-167
Change-Id: Ic01d199a798626088ed002577e21e4d0c559ee16
Signed-off-by: Allain Legacy <alegacy@redhat.com>