Remove assumptions about vendor values
This removes assumptions made about vendor-specific contents in
description and extension attributes. The specification documents
do not define the contents of extension or description fields;
therefore, the tests should not look for any specific values beyond
checking for their presence.

It also removes tests steps related to specific types of resources
that are not defined in the test specification document. Specifically,
7.3.2 steps s14 to s18 are not a part of the test specification
document

Issue-ID: INT-168
Change-Id: I7e1686c7487ce7b2fd75e303f88d41b3b6c48cc5
Signed-off-by: Allain Legacy <alegacy@redhat.com>